requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.26.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
